Revert "Switch to Quad114 as the default resolver"
This reverts commit 91f97833a3a97c097fb75b7017e6b67a4cb2d7d4.

The Internet has become a sad place.

People in China need to use resolvers in China.
People in the US would not trust resolvers in China.
People in the EU would not trust resolvers in the US.

Revert to Quad9 for now, and add some documentation about why
that might be changed (especially in China) later.