[CI] Leverage `after_success` phase in .travis.yml
There were two `ls -l dnscrypt-proxy-*.tar.gz dnscrypt-proxy-*.zip`
commands in `.travis.yml`, one in the `script`(build) phase, one in the
`after_deploy` phase, they were actually duplicated.

As deployment job won't create any new files, and the command isn't part
of the steps to build artifacts, but to confirm/list the built files,
this action looks more suitable to be placed in the `after_success`, to
be separated with the build commands, and the duplicated one in the
`after_deploy` phase could be removed.