plugin_block_name: also block cname
Currently, CNAME can be used to resolve blocked domain (Just CNAME an
allowed domain to a blocked one and the server often return both the
CNAME and the A or AAAA field if it exists).

This is a naive implementation that check if an answer exists, check
every CNAME field in it against the blacklist.
It stops on the first error or on the first blocked CNAME.